isafw: resolved the issue caused by a bug in cve-check-tool of database update
Last commit has the change to get the source code, fetching cve feeds from
year 2010. It was not the right thing to do becuase issue was not with the
meta files but with the format of SHAsum of the data. NVD was using SHAsum in
lower case but now changed it to the upper case causing issue in Cve-check-tool.
cve-check-tool made the changes in dev code but in the release so now using
temporarily thier dev code to get the cve report.

Signed-off-by: NaeemKK <naeem_khan@mentor.com>