isafw: changes for mapping of kernel names and solution for cve-check-tool issue
NVD has changed the way it used to provide hash of their database, which caused
issues with the version used for cve-check-tool. So we are now fetching upto a commit
which solves these problems.Another change is for kernel name mapping as we are using
our custom names for kernel.

Signed-off-by: NaeemKK <naeem_khan@mentor.com>